Add token retrieval.
we orchestrate the retrieval of the token in the same way as we would do in production.
We call getToken that either goes looking for the token in localStorage or uses the memorized value.
As a predicted usage I have set it in all of the constructors for our external services.
This way whenever we make a fetch request we can reference this.token.