Reset and variables.
To maintain our lean way of living in terms of CSS we need to make sume assumptions.

We change from extending the reset class in our .scss to adding ddb-reset as a class to the elements we want to reset.

We also introduce a design-system that does not export any style blocks by itself but only exports variables and mixins.