Expand button atom.
We need to take a lot more into account when a button can be included in a form element. Which it should most of the time.

Besides that we have tried to mirror the styling on the actual ddb site even more. It won't fit 100% before we can see it side by side.