Support polyfill of deps.
Babel won't touch the node_modules unless we make use of babel.config.js instead of .babelrc.
    https://babeljs.io/docs/en/configuration#whats-your-use-case

Then WebPack need to still exclude all of the other packages since we still will be handling this on a need to know basis.

Then when we do now that we are making use of a package that is not IE11 compatible we can add that package name to polyfill.config.js in order to include it into the transpilation pipeline.

Including node_modules into your transpilation pipeline is not unheard of. create-react-app does it as well to ensure compatability.