Rework release process
The current split between NPM packages and GitHub releases creates
some complexity in the process with if-statements and reliance on
commits and tags being pushed simultaniously.

Instead we split the process differently:

- One workflow updates the latest GitHub release used by DDB CMS from
the master branch
- One workflow builds GitHub releases and NPM packages based on tags

The build process is also optimized using more dedicated actions:

- bahmutov/npm-install provides dependency installation with cache
properly configured in a single step. We trust this action since the
developer is also leading the Cypress project which we use.

We intentionally keep using GHR. Other actions will not
delete/recreate existing releases and thus Latest release will be 
pushed to the bottom. For consistency we use it for all types of
GitHub releases.

Finally tag release process is split into a single build step passing
the results to separate jobs for publishing new relase and NPM
package.