Set package name and repository based on current organisation
Replace reload with danskernesdigitale bibliotek in package name and
repository. This matches the current origin of the project.

Data in package.json must match current repository for NPM package
publication to work. Consequently we update package name and
repository to match the current repository.

This allows publication to local NPM repositories for forks.