Use fan-out architecture when building pull requests
Fetching dependencies is by far the most time consuming proces when
building. With this approach we only have to do it once.

Then we can let each tool download the result as an artifact and run
its own course.

bahmutov/npm-install is simpler than having to manage cache ourselves
and built by a developer we trust since he is also the lead of
Cypress.