Update linting and formatting setup.
Main objective of these changes is remove some of the friction from when
working on the project.

Before we had eslint and prettier intertwined which made for a "secure"
experience but made it so that formatting errors that were auto-fixable
were being treated as errors. This would make the editing experience
slow and tiresome when you would not have the appropriate editor setup
that would auto-fix everything for you.

These changes seperated eslint and prettier and stylelint and prettier.
Letting eslint and stylelint be responsible for code quality and our own
weird formatting rules whereas prettiers concerns is 100% on auto-fixable
formatting.

This results in no more red lines in our IDE's when it comes to prettier
warnings which eases development but still maintains security of formatted
code due to github actions and the proposed IDE configs supplied.

Besides the seperation we also bump eslint and prettier a major version.